fixed flutter lint error.
