+++ example/pubspec.yaml
@@ -29,6 +29,8 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.8
 
+  audio_session: any
+  permission_handler: any
 dev_dependencies:
   integration_test:
     sdk: flutter
