+++ test/haishin_kit_method_channel_test.dart
@@ -6,10 +6,11 @@ void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
 
   MethodChannelHaishinKit platform = MethodChannelHaishinKit();
-  const MethodChannel channel = MethodChannel('haishin_kit');
+  const MethodChannel channel = MethodChannel('com.haishinkit');
 
   setUp(() {
-    TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger.setMockMethodCallHandler(
+    TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
+        .setMockMethodCallHandler(
       channel,
       (MethodCall methodCall) async {
         return '42';
@@ -18,7 +19,8 @@ void main() {
   });
 
   tearDown(() {
-    TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger.setMockMethodCallHandler(channel, null);
+    TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
+        .setMockMethodCallHandler(channel, null);
   });
 
   test('getPlatformVersion', () async {
